Refactor: Unfilted cards are now passed to category and filtered within
Previously, filtered results are passed to category and is displayed.
However, this prevents the category scene from refreshing when cards are
deleted since the cards state is in the home scene, a sibling of
category scene.

Additional refactors because of this:
searchAll function is now responsible for filtering both category and
search within the category scene since category scene is displaying both
search and home category redirects. Route parameters to the category have
to be changed for both category and search redirects to accomadate this.