Feat: users can delete tags/ingredients input field when creating card
Tags and ingredients are now stored as objects instead of arrays in the
form state of create new card. Previously, input fields are identified
from it's neighbour by its index in the array. Since the index of a
input field will change when the array it belongs in is mutated, feature
such as deleting an inputfield cannot be supported. By labeling each
input field with a unique id in a key value pair format, each field can
be referenced reliably.