Fix: new cards are saved in the wrong format on localstorage
Tags and ingredients are supposed to be saved as arrays under cards object on local storage.
Previously, when saving new cards, tags and ingredients were erroneously saved as objects, with the key being an unique id. Thus, causing an error when the searchAll function treats tags/ingredients property of the cards object as an array.